# docs/swagger.yaml
basePath: /api
definitions:
  dto.ChangeRoleRequest:
    properties:
      NewRole:
        type: integer
      id:
        type: integer
    type: object
...
  dto.CreateStatusRequest:
    properties:
      status_name:
        type: string
    type: object
  dto.CreateStatusResponse:
    properties:
      id:
        type: integer
    type: object
  dto.CreateTypeRequest:
    properties:
      type_name:
        type: string
    type: object
  dto.CreateTypeResponse:
    properties:
      id:
        type: integer
    type: object
...
  dto.GetStatusByIdResponse:
    properties:
      id:
        type: integer
      status_name:
        type: string
    type: object
  dto.GetTypeByIdResponse:
    properties:
      id:
        type: integer
      type_name:
        type: string
    type: object
  dto.GetUserByIdResponse:
    properties:
      email:
        type: string
      first_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
      user_password:
        type: string
      user_role:
        type: integer
      username:
        type: string
    type: object
...
  dto.IncidentStatus:
    properties:
      id:
        type: integer
      status_name:
        type: string
    type: object
  dto.IncidentType:
    properties:
      id:
        type: integer
      type_name:
        type: string
    type: object
...
  dto.UpdateStatusRequest:
    properties:
      id:
        type: integer
      status_name:
        type: string
    type: object
  dto.UpdateTypeRequest:
    properties:
      id:
        type: integer
      type_name:
        type: string
    type: object
  dto.UpdateUserRequest:
    properties:
      email:
        type: string
      first_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
      user_password:
        type: string
      user_role:
        type: integer
      username:
        type: string
    type: object
  dto.User:
    properties:
      email:
        type: string
      first_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
      user_password:
        type: string
      user_role:
        type: integer
      username:
        type: string
    type: object
...
  ginhttp.response:
    properties:
      message:
        type: string
    type: object
info:
  contact:
    email: ekaterinaamzr@gmail.com
  description: Swagger API for Golang Project Green Alarm.
  license:
    name: MIT
    url: https://github.com/MartinHeinz/go-project-blueprint/blob/master/LICENSE
  termsOfService: http://swagger.io/terms/
  title: Green Alarm
  version: "1.0"
paths:
...
  /statuses:
    get:
      description: Get list of statuses
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.IncidentStatus'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ginhttp.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ginhttp.response'
      security:
      - ApiKeyAuth: []
      summary: Get all
      tags:
      - Statuses
    post:
      consumes:
      - application/json
      description: Create status
      parameters:
      - description: New status data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.CreateStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CreateStatusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ginhttp.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ginhttp.response'
      security:
      - ApiKeyAuth: []
      summary: Create
      tags:
      - Statuses
  /statuses/{id}:
    delete:
      description: Delete status
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ginhttp.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ginhttp.response'
      security:
      - ApiKeyAuth: []
      summary: Delete
      tags:
      - Statuses
    get:
      description: Get status by id
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.GetStatusByIdResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ginhttp.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ginhttp.response'
      security:
      - ApiKeyAuth: []
      summary: Get by id
      tags:
      - Statuses
    put:
      consumes:
      - application/json
      description: Update status
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      - description: Updated status data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateStatusRequest'
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ginhttp.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ginhttp.response'
      security:
      - ApiKeyAuth: []
      summary: Update
      tags:
      - Statuses
  /types:
    get:
      description: Get list of types
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.IncidentType'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ginhttp.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ginhttp.response'
      security:
      - ApiKeyAuth: []
      summary: Get all
      tags:
      - Types
    post:
      consumes:
      - application/json
      description: Create type
      parameters:
      - description: New type data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.CreateTypeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CreateTypeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ginhttp.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ginhttp.response'
      security:
      - ApiKeyAuth: []
      summary: Create
      tags:
      - Types
  /types/{id}:
    delete:
      description: Delete type
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ginhttp.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ginhttp.response'
      security:
      - ApiKeyAuth: []
      summary: Delete
      tags:
      - Types
    get:
      description: Get type by id
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.GetTypeByIdResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ginhttp.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ginhttp.response'
      security:
      - ApiKeyAuth: []
      summary: Get by id
      tags:
      - Types
    put:
      consumes:
      - application/json
      description: Update type
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      - description: Updated type data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateTypeRequest'
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ginhttp.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ginhttp.response'
      security:
      - ApiKeyAuth: []
      summary: Update
      tags:
      - Types
  /users:
    get:
      description: Get list of users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.User'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ginhttp.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ginhttp.response'
      security:
      - ApiKeyAuth: []
      summary: Get all
      tags:
      - Users
  /users/{id}:
    delete:
      description: Delete user
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ginhttp.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ginhttp.response'
      security:
      - ApiKeyAuth: []
      summary: Delete
      tags:
      - Users
    get:
      description: Get user by id
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.GetUserByIdResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ginhttp.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ginhttp.response'
      security:
      - ApiKeyAuth: []
      summary: Get by id
      tags:
      - Users
    patch:
      consumes:
      - application/json
      description: Change user role
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      - description: Updated user role
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.ChangeRoleRequest'
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ginhttp.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ginhttp.response'
      security:
      - ApiKeyAuth: []
      summary: Change role
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Update user
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      - description: Updated user data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateUserRequest'
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ginhttp.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ginhttp.response'
      security:
      - ApiKeyAuth: []
      summary: Update
      tags:
      - Users
securityDefinitions:
  ApiKeyAuth:
    in: Bearer token
    name: Authorization
    type: apiKey
swagger: "2.0"